Remove Winston from github logger to reduce dependency size
